moved all prompts to 1 file
