--- mdagent/subagents/prompts/critic_prompts.py
@@ -1,85 +1,0 @@
-# code critic
-code_critic_format = """
-You should only respond in JSON format as described below:
-{{
-    "code_quality": "code_quality",
-    "output_validity": "output_validity",
-    "success": boolean,
-    "critique": "critique",
-    "suggestions": "suggestions"
-}}
-Ensure the response can be parsed by Python `json.loads`,
-e.g.: no trailing commas, no single quotes, etc.
-"""
-
-code_critic_prefix = """
-You are an assistant that assesses the quality of my code
-and the validity of its output for my molecular dynamics project,
-and provides useful guidance.
-
-You are required to evaluate the quality of my code,
-the correctness of its output and if it meets the task requirements.
-Exceeding the requirements is also considered a success,
-while failing to meet them requires you to
-provide critique and suggestions to help me improve.
-
-I will give you the following information.
-
-Code: The source code
-Code Output: The output of the executed code (if applicable, including execution errors)
-Task: The objective that the code needs to accomplish
-Context: The context of the task (optional, if applicable)
-"""
-
-code_critic_prompt = """
-INPUT:
-code: {code},
-output: {code_output},
-task: {task},
-context: {context}
-"""
-
-# task critic
-task_critic_format = """
-You should only respond in JSON format as described below:
-{{
-    "success": boolean,
-    "critique": string,
-}}
-Ensure the response can be parsed by Python `json.loads`,
-e.g.: no trailing commas, no single quotes, etc.
-"""
-
-task_critic_prefix = """
-You are an assistant that assesses my progress on my molecular
-dynamics project and provides useful guidance.
-
-You are required to evaluate if I have succesfully met
-all of the task requirements.
-If I have met or exceeded the requirements,
-your response should be True.
-However, if I have failed to meet the requirements,
-your response should be False,
-accompanied by critique to help me improve.
-
-I will give you the following information.
-
-
-Files: All files created in the current directory,
-along with descriptions when applicable.
-Code: The code I have written
-Output: The execution output of the code I have written
-Task: The objective I need to accomplish
-Context: The context of the task
-Additional Information: Any additional information, in case you cause an error.
-"""
-
-task_critic_prompt = """
-INPUT:
-files: {files},
-code: {code},
-output: {code_output},
-task: {task},
-context: {context},
-additional_information: {additional_information}
-"""
